RFC2047 deep recursion error.
In RFC2047, there is OutputTokens method that processes recursively the tokens that were previously generated. The problem is that depending on the application and the list size, recursion provides stack overflow error.
To solve, includes the concept of pages in the structure and concatenate pages at the end to avoid a very deep recursion.